requirements.txt
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
decorator==4.4.2
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.31.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==1.0.3
numpy==2.3.3
oauthlib==3.3.1
pillow==11.3.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.4
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
tqdm==4.67.1
uritemplate==4.2.0
urllib3==2.5.0
